Navigation drawer is fully functional. It doesn't close when an item is selected. Doesn't work from wishlists menu either.
