+++ app/src/main/java/com/daviancorp/android/ui/general/GenericActivity.java
@@ -33,156 +33,26 @@ import com.daviancorp.android.ui.dialog.AboutDialogFragment;
  *  - override createFragment() for detail fragments
  */
 
-public abstract class GenericActivity extends ActionBarActivity {
-
-	protected static final String DIALOG_ABOUT = "about";
-
-	protected Fragment detail;
-    private ListView mDrawerList;
-    private ArrayAdapter<String> mAdapter;
-    public ActionBarDrawerToggle mDrawerToggle;
-    public DrawerLayout mDrawerLayout;
+public abstract class GenericActivity extends GenericActionBarActivity {
 
 	protected abstract Fragment createFragment();
-	
-	@Override
-	public void onCreate(Bundle savedInstanceState) {
-		super.onCreate(savedInstanceState);
-		setContentView(R.layout.activity_main);
 
-        mDrawerLayout = (DrawerLayout)findViewById(R.id.drawer_layout);
-        // Populate navigation drawer
-        mDrawerList = (ListView)findViewById(R.id.navList);
-        addDrawerItems();
-        mDrawerList.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                // Set navigation actions
-                //
-                //
-                //
-                //
-            }
-        });
-        setupDrawer();
-
-		FragmentManager fm = getSupportFragmentManager();
-		Fragment fragment = fm.findFragmentById(R.id.fragment_container);
-
-		if (fragment == null) {
-			fragment = createFragment();
-			fm.beginTransaction().add(R.id.fragment_container, fragment)
-					.commit();
-		}
-		
-		setTitle(R.string.app_name);
-		getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-		getSupportActionBar().setHomeButtonEnabled(true);
-	}
-
-    // Set up drawer menu options
-    private void addDrawerItems(){
-        String[] menuArray = getResources().getStringArray(R.array.drawer_items);
-        mAdapter = new ArrayAdapter<String>(this, android.R.layout.simple_list_item_1, menuArray);
-        mDrawerList.setAdapter(mAdapter);
-    }
-
-
-    // Set up drawer toggle actions
-    private void setupDrawer(){
-        mDrawerToggle = new ActionBarDrawerToggle(this, mDrawerLayout, R.string.drawer_open, R.string.drawer_close){
-            /** Called when a drawer has settled in a completely open state. */
-            public void onDrawerOpened(View drawerView){
-                super.onDrawerOpened(drawerView);
-                invalidateOptionsMenu(); // Creates call to onPrepareOptionsMenu()
-            }
-            /** Called when a drawer has settled in a completely closed state. */
-            public void onDrawerClosed(View view){
-                super.onDrawerClosed(view);
-                invalidateOptionsMenu(); // Creates call to onPrepareOptionsMenu()
-            }
-        };
-
-        // Enable menu button to toggle drawer
-        mDrawerToggle.setDrawerIndicatorEnabled(false);
-        mDrawerLayout.setDrawerListener(mDrawerToggle);
-    }
-
-    public void enableDrawerIndicator(){
-        mDrawerToggle.setDrawerIndicatorEnabled(true);
-    }
-
-    // Sync button animation sync with drawer state
     @Override
-    protected void onPostCreate(Bundle savedInstanceState){
-        super.onPostCreate(savedInstanceState);
-        mDrawerToggle.syncState();
-    }
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+        super.setupDrawer(); // Needs to be called after setContentView
+        FragmentManager fm = getSupportFragmentManager();
+        Fragment fragment = fm.findFragmentById(R.id.fragment_container);
 
-    // Handle toggle state sync across configuration changes (rotation)
-    @Override
-    public void onConfigurationChanged(Configuration newConfig){
-        super.onConfigurationChanged(newConfig);
-        mDrawerToggle.onConfigurationChanged(newConfig);
-    }
-
-	@Override
-	public boolean onOptionsItemSelected(MenuItem item) {
-
-        // Detect navigation drawer item selected
-        if(mDrawerToggle.onOptionsItemSelected(item)){
-            return true;
+        if (fragment == null) {
+            fragment = createFragment();
+            fm.beginTransaction().add(R.id.fragment_container, fragment)
+                    .commit();
         }
 
-        // Detect home and or expansion menu item selections
-		switch (item.getItemId()) {
-		case android.R.id.home:
-            // For the love of god fix this
-			// Whenever the home button is pressed, go back to home and clear the stack of activities
-			Intent intent = new Intent(GenericActivity.this, HomeActivity.class);
-			intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
-			startActivity(intent);
-			return true;
-		case R.id.about:
-			FragmentManager fm = getSupportFragmentManager();
-			AboutDialogFragment dialog = new AboutDialogFragment();
-			dialog.show(fm, DIALOG_ABOUT);
-			return true;
-
-        case R.id.send_feedback:
-            Intent email = new Intent(Intent.ACTION_SEND);
-            email.setType("text/email");
-            email.putExtra(Intent.EXTRA_EMAIL, new String[]{"monster-hunter-database-feedback@googlegroups.com"});
-            email.putExtra(Intent.EXTRA_SUBJECT, "MH4U Database Feedback");
-            startActivity(Intent.createChooser(email, "Send Feedback:"));
-
-		default:
-			return super.onOptionsItemSelected(item);
-		}
-	}
-
-	@Override
-	public boolean onCreateOptionsMenu(Menu menu) {
-		super.onCreateOptionsMenu(menu);
-		MenuInflater inflater = getMenuInflater();
-		inflater.inflate(R.menu.main, menu);
-		return true;
-	}
-
-	@Override
-	public boolean onPrepareOptionsMenu(Menu menu) {
-		for (int i = 0; i < menu.size(); i++) {
-			MenuItem mi = menu.getItem(i);
-			String title = mi.getTitle().toString();
-			Spannable newTitle = new SpannableString(title);
-			newTitle.setSpan(new ForegroundColorSpan(Color.WHITE), 0,
-					newTitle.length(), Spannable.SPAN_EXCLUSIVE_EXCLUSIVE);
-			mi.setTitle(newTitle);
-		}
-		return true;
-	}
-
-	public Fragment getDetail() {
-		return detail;
-	}
+        setTitle(R.string.app_name);
+        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+        getSupportActionBar().setHomeButtonEnabled(true);
+    }
 }
