+++ app/src/main/java/com/daviancorp/android/ui/general/GenericTabActivity.java
@@ -24,9 +24,7 @@ import com.daviancorp.android.ui.dialog.AboutDialogFragment;
  *  - override onCreate() to set title
  */
 
-public abstract class GenericTabActivity extends ActionBarActivity{
-
-	protected static final String DIALOG_ABOUT = "about";
+public abstract class GenericTabActivity extends GenericActionBarActivity{
 
 	protected Fragment detail;
 
@@ -40,52 +38,6 @@ public abstract class GenericTabActivity extends ActionBarActivity{
 		getSupportActionBar().setHomeButtonEnabled(true);
 
 		setContentView(R.layout.activity_tab);
-
-	}
-
-	@Override
-	public boolean onOptionsItemSelected(MenuItem item) {
-		switch (item.getItemId()) {
-		case android.R.id.home:
-			// Whenever the home button is pressed, go back to home and clear the stack of activities
-			Intent intent = new Intent(GenericTabActivity.this,
-					HomeActivity.class);
-			intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
-			startActivity(intent);
-			return true;
-		case R.id.about:
-			FragmentManager fm = getSupportFragmentManager();
-			AboutDialogFragment dialog = new AboutDialogFragment();
-			dialog.show(fm, DIALOG_ABOUT);
-
-			return true;
-		default:
-			return super.onOptionsItemSelected(item);
-		}
-	}
-
-	@Override
-	public boolean onCreateOptionsMenu(Menu menu) {
-		super.onCreateOptionsMenu(menu);
-		MenuInflater inflater = new MenuInflater(getApplicationContext());
-		inflater.inflate(R.menu.main, menu);
-		return true;
-	}
-
-	@Override
-	public boolean onPrepareOptionsMenu(Menu menu) {
-		for (int i = 0; i < menu.size(); i++) {
-			MenuItem mi = menu.getItem(i);
-			String title = mi.getTitle().toString();
-			Spannable newTitle = new SpannableString(title);
-			newTitle.setSpan(new ForegroundColorSpan(Color.WHITE), 0,
-					newTitle.length(), Spannable.SPAN_EXCLUSIVE_EXCLUSIVE);
-			mi.setTitle(newTitle);
-		}
-		return true;
-	}
-
-	public Fragment getDetail() {
-		return detail;
+        setupDrawer(); // Needs to be called after setContentView
 	}
 }
